Extract shared TransformPrivateState to unify simple and carousel item state machines
Both models share the same 4-phase state machine (tracking/inertia/snapping/settled).
TransformConfig provides optional bounds clamping and snap target for both use cases.
