+++ packages/core/src/model/index.ts
@@ -1,3 +1,4 @@
 export * from "./primitives.js";
+export * from "./transform.js";
 export * from "./simple.js";
 export * from "./carousel.js";
